+++ silkworm/db/state/account_codecs.hpp
@@ -14,21 +14,29 @@
    limitations under the License.
 */
 
-#include "address_decoder.hpp"
+#pragma once
 
-#include <catch2/catch_test_macros.hpp>
+#include <silkworm/core/types/account.hpp>
+#include <silkworm/db/datastore/snapshots/common/codec.hpp>
+#include <silkworm/infra/common/decoding_exception.hpp>
 
-#include <silkworm/core/common/util.hpp>
+#include "silkworm/db/util.hpp"
 
 namespace silkworm::db::state {
 
-TEST_CASE("AddressDecoder") {
-    using evmc::literals::operator""_address;
-    AddressDecoder decoder;
-    decoder.decode_word(*from_hex("0x000000000000000000636f6e736f6c652e6c6f67"));
-    CHECK(decoder.value == 0x000000000000000000636f6e736f6c652e6c6f67_address);
+struct AccountDecoder : public snapshots::Decoder {
+    Account value;
 
-    CHECK_THROWS_AS(decoder.decode_word({}), std::runtime_error);
-}
+    ~AccountDecoder() override = default;
+
+    void decode_word(ByteView word) override {
+        auto account = Account::from_encoded_storage_v3(word);
+        if (!account)
+            throw DecodingException{account.error(), "AccountDecoder failed to decode Account"};
+        value = std::move(*account);
+    }
+};
+
+static_assert(snapshots::DecoderConcept<AccountDecoder>);
 
 }  // namespace silkworm::db::state
