+++ silkworm/db/state/log_address_inverted_index.hpp
@@ -19,7 +19,7 @@
 #include <silkworm/db/datastore/snapshots/common/raw_codec.hpp>
 #include <silkworm/db/datastore/snapshots/segment/kv_segment_reader.hpp>
 
-#include "address_decoder.hpp"
+#include "address_codecs.hpp"
 
 namespace silkworm::db::state {
