+++ silkworm/db/state/storage_domain.hpp
@@ -16,41 +16,12 @@
 
 #pragma once
 
-#include <stdexcept>
-
+#include <silkworm/db/datastore/kvdb/domain.hpp>
 #include <silkworm/db/datastore/snapshots/segment/kv_segment_reader.hpp>
 
-#include "address_decoder.hpp"
+#include "storage_codecs.hpp"
 
 namespace silkworm::db::state {
-
-struct Bytes32Decoder : public snapshots::Decoder {
-    evmc::bytes32 value;
-    ~Bytes32Decoder() override = default;
-    void decode_word(ByteView word) override {
-        if (word.size() < sizeof(value.bytes))
-            throw std::runtime_error{"Bytes32Decoder failed to decode"};
-        std::memcpy(value.bytes, word.data(), sizeof(value.bytes));
-    }
-};
-
-static_assert(snapshots::DecoderConcept<Bytes32Decoder>);
-
-struct StorageAddressAndLocationDecoder : public snapshots::Decoder {
-    struct {
-        AddressDecoder address;
-        Bytes32Decoder location_hash;
-    } value;
-
-    ~StorageAddressAndLocationDecoder() override = default;
-
-    void decode_word(ByteView word) override {
-        value.address.decode_word(word);
-        value.location_hash.decode_word(word.substr(kAddressLength));
-    }
-};
-
-static_assert(snapshots::DecoderConcept<StorageAddressAndLocationDecoder>);
 
 using StorageDomainKVSegmentReader = snapshots::segment::KVSegmentReader<StorageAddressAndLocationDecoder, Bytes32Decoder>;
